refactor(core): move image preprocessing into model adapter
