+++ packages/core/src/ai-model/model-adapter/image-preprocess.ts
