+++ packages/core/src/ai-model/model-adapter/types.ts
@@ -17,10 +17,10 @@ import type {
   LocateResultAdapterDefinition,
   ResolvedLocateResultCoordinates,
 } from '../shared/model-locate-result/types';
-import type { ImagePreprocessPolicy } from '../workflows/image-preprocess';
 import type { LocateFn } from '../workflows/inspect/types';
 import type { PlanFn } from '../workflows/planning/types';
 import type { CustomPlanningDefinition } from './custom-planning-types';
+import type { ImagePreprocessPolicy } from './image-preprocess';
 
 export type {
   ImagePreprocessPolicy,
